Create README.md
Hi, 
Anders here from PARAGAIA!
We are building a new national data platform and will soon need to include new members to our dev team! Initial we are trying to find 2 - 4 Full Stack Developers, and one Data Engineer. 
Technologies used will be Postgress, MongoDB, Python, Spark, C#/ Core, OpenShift to name a few. 
I will come back with more specific requests as we get closer to the start of the project. If you are curious about this and have the opportunity to work in the Stockholm area, please get in touch and I will tell you more!